* Close #3 - Linux Porting
* makefile 및 make 스크립트 추가
* 빌드 오류 수정
* 환경 변경으로 인한 버그 수정

+++ src/log/SCLogger.cpp
@@ -58,7 +58,7 @@ void SCLogger::setLevel(ELogLevel logLevel)
 
 void SCLogger::log(ELogLevel logLevel, const SCChar* format, ...)
 {
-	if (logLevel > _logLevel)
+	if ((int)logLevel > (int)_logLevel)
 	{
 		return;
 	}
